feat(pipeline): permafrost

+++ pipeline/dags/task_factories.py
@@ -32,7 +32,7 @@ def gcs_download_file(bucket_name: str, dir: str, appendix: str = ''):
         return local_filename
     return fn
 
-def gcs_upload_file(bucket_name=str, layer_id=str, layer_variable=str, layer_version=str):
+def gcs_upload_file(bucket_name: str, layer_id: str, layer_variable: str, layer_version: str):
     @task(task_id='gcs_upload_file')
     def fn(filename: str):
         hook = GCSHook('google')
@@ -40,6 +40,18 @@ def gcs_upload_file(bucket_name=str, layer_id=str, layer_variable=str, layer_ver
         hook.upload(bucket_name, remote_filename, filename)
         return filename
     return fn
+
+def gcloud_upload_dir(bucket_name: str, layer_id: str, layer_variable: str, layer_version: str, directory: str):
+    return BashOperator(
+        task_id='gcloud_upload',
+        bash_command='gcloud auth activate-service-account --key-file $KEY_FILE && gsutil -m cp -r $UPLOAD_DIR/* $BUCKET',
+        env={
+            "UPLOAD_DIR": directory,
+            "BUCKET": f'gs://{bucket_name}/{layer_id}.{layer_variable}/{layer_version}/final/',
+            "KEY_FILE": '/opt/airflow/plugins/service-account.json',
+            "CLOUDSDK_PYTHON": '/usr/local/bin/python'
+        }
+    )
 
 def gdal_info():
     return BashOperator(
@@ -49,56 +61,65 @@ def gdal_info():
     )
 
 def legend_image(bucket_name: str, layer_id: str, layer_variable: str, layer_version: str, workdir: str, color_file: str):
-    @task_group(group_id='legend_image_group')
-    def fn():
-        upload_legend = gcs_upload_file(bucket_name=bucket_name, layer_id=layer_id, layer_variable=layer_variable, layer_version=layer_version)
-        legend_image = BashOperator(
-            task_id='legend_image',
-            bash_command=f'rm -f $FILEPATH_OUT && node /opt/airflow/plugins/generate-legend-image.js && echo $FILEPATH_OUT',
-            env={"FILEPATH_OUT": f'{workdir}/legend.png', "COLOR_FILE": color_file}
-        )
-        legend_image >> upload_legend(legend_image.output)
-    return fn
+    return BashOperator(
+        task_id='legend_image',
+        bash_command=f'rm -f $FILEPATH_OUT && node /opt/airflow/plugins/generate-legend-image.js && echo $FILEPATH_OUT',
+        env={"FILEPATH_OUT": f'{workdir}/legend.png', "COLOR_FILE": color_file}
+    )
 
 # generate timestamps array from filenames and write metadata to json file then upload to gcs
 def metadata(bucket_name: str, layer_id: str, layer_variable: str, layer_version: str, workdir: str, metadata: dict):
-    @task_group(group_id='metadata_group')
+    @task(task_id='metadata')
     def fn(files):
-        @task()
-        def write_metadata_json(files):
-            timestamps = str(files)
-            extended_metadata = dict({"timestamps": timestamps}, **metadata)
-            filepath = str(Path(workdir).joinpath('metadata.json'))
+        timestamps = list(map(helper.filename_to_date, files))
+        extended_metadata = dict(metadata, **{"timestamps": timestamps})
+        filepath = str(Path(workdir).joinpath('metadata.json'))
 
-            with open(filepath, "w") as f:
-                metadata_string = json.dumps(extended_metadata, indent=4)
-                f.write(metadata_string)
-                print(metadata_string)
+        with open(filepath, "w") as f:
+            metadata_string = json.dumps(extended_metadata, indent=4)
+            f.write(metadata_string)
+            print(metadata_string)
 
-            return filepath
-
-        upload_metadata = gcs_upload_file(bucket_name=bucket_name, layer_id=layer_id, layer_variable=layer_variable, layer_version=layer_version)
-
-        return upload_metadata(write_metadata_json(files))
+        return filepath
     return fn
 
-def gdal_transforms(layer_variable: str, color_file: str, gdal_te: str = '-180 -90 180 90', gdal_ts: str = '1024 512'):
+def gdal_transforms(layer_variable: str, color_file: str, layer_type: str, zoom_levels: str, gdal_te: str = '-180 -90 180 90', gdal_ts: str = '1024 512', warp_cmd: str = None, max_tis_warp: int = 4,  max_tis_dem: int = 4, max_tis_translate: int = 4):
+    def get_transform_task():
+        if layer_type == 'image':
+            return BashOperator.partial(
+                task_id='gdal_translate_image',
+                bash_command='rm -f $FILEPATH_OUT && gdal_translate -of PNG $FILEPATH_IN $FILEPATH_OUT && echo $FILEPATH_OUT',
+                max_active_tis_per_dag=max_tis_translate
+            )
+
+        return BashOperator.partial(
+            task_id='gdal_translate_tiles',
+            bash_command=f'rm -rf $FILEPATH_OUT && gdal2tiles.py --profile geodetic --zoom={zoom_levels} --tmscompatible --no-kml --webviewer=none --resampling average --s_srs EPSG:4326 $FILEPATH_IN $FILEPATH_OUT && echo $FILEPATH_OUT' ,
+            max_active_tis_per_dag=max_tis_translate
+        )
+
+    def get_transform_outpath(filename):
+        if layer_type == 'image':
+            return helper.change_filename(filename, appendix='', suffix='png')
+
+        return helper.change_filename(filename, appendix='', remove_suffix=True)
+
     @task_group(group_id='gdal_transforms_group')
     def fn(downloads):
+        warp_command = f'gdalwarp -t_srs EPSG:4326 -te {gdal_te} -ts {gdal_ts} -r near --config GDAL_CACHEMAX 90% -co compress=LZW NETCDF:"$FILEPATH_IN":$DATA_VARIABLE $FILEPATH_OUT' if not warp_cmd else warp_cmd
         gdal_warp = BashOperator.partial(
             task_id='gdal_warp',
-            bash_command=f'rm -f $FILEPATH_OUT && gdalwarp -t_srs EPSG:4326 -te {gdal_te} -ts {gdal_ts} -r near --config GDAL_CACHEMAX 90% -co compress=LZW NETCDF:"$FILEPATH_IN":$DATA_VARIABLE $FILEPATH_OUT && echo $FILEPATH_OUT'
+            bash_command=f'rm -f $FILEPATH_OUT && {warp_command} && echo $FILEPATH_OUT',
+            max_active_tis_per_dag=max_tis_warp
         )
 
         gdal_dem = BashOperator.partial(
             task_id='gdal_dem',
-            bash_command='rm -f $FILEPATH_OUT && gdaldem color-relief $FILEPATH_IN $COLOR_FILE --config GDAL_CACHEMAX 90% -co compress=LZW -alpha $FILEPATH_OUT && echo $FILEPATH_OUT'
+            bash_command='rm -f $FILEPATH_OUT && gdaldem color-relief $FILEPATH_IN $COLOR_FILE --config GDAL_CACHEMAX 90% -co compress=LZW -alpha $FILEPATH_OUT && echo $FILEPATH_OUT',
+            max_active_tis_per_dag=max_tis_dem
         )
 
-        gdal_translate = BashOperator.partial(
-            task_id='gdal_translate',
-            bash_command='rm -f $FILEPATH_OUT && gdal_translate -of PNG $FILEPATH_IN $FILEPATH_OUT && echo $FILEPATH_OUT'
-        )
+        gdal_translate = get_transform_task()
 
         gdal_warps = gdal_warp.expand(
             env=downloads.map(lambda filename: {
@@ -119,7 +140,7 @@ def gdal_transforms(layer_variable: str, color_file: str, gdal_te: str = '-180 -
         gdal_translates = gdal_translate.expand(
             env=gdal_dems.output.map(lambda filename: {
                 "FILEPATH_IN": filename,
-                "FILEPATH_OUT": helper.change_filename(filename, appendix='', suffix='png'),
+                "FILEPATH_OUT": get_transform_outpath(filename)
             })
         )
